agent_app/requirements.txt
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
Booktype==1.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-scheduler==0.10.1
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
icalendar==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psutil==6.1.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
redis>=4.0.0
requests==2.32.3
setuptools==75.4.0
simple-websocket==1.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
